Add forward, back, left, right, yaw, and pitch inputs which are exposed as DataElements on the component so they have text fields that can define each key within the editor
Use the StartingPointInput gem and its InputEventNotificationBus to
track when an input is pressed, released, or held.

Each of the input names are stored in separate AZStd::string variables
because exposing the StartingPointInput::InputNotificationId variables directly via the DataElement call does not present a text field inside the editor. So instead, strings are used and then the Input IDs are assigned to their values using those strings. The keys are assigned when the entity is activated.