Make sure that the sprint key modifier still applies if pressed before the forward/left/right movement keys
Make it so that the sprint key is decelerated (decreased) when either
none of the forward/left/right keys are pressed or when only moving
backwards (accounting for if both left & right are simultaneously
pressed).

Remove the conditional statement that checks if the back key is pressed
to apply the sprint factor to the final velocity vector assignment.
Previously having this condition created a potential jump discontinuity
in the velocity when moving backwards and left/right in a diagonal and
then releasing either the left or right key (immediately removing the
sprint factor when only back was pressed). The current solution is to
simply decelerate the sprint multiplier when only back is held.