Shave off the extra ramp time for sprint
total_ramp_time was directly using the event value multiplier as a
factor in its calculation. Which makes sense for the movement keys which
are either 0 when released or some non-zero value (like 1.0) when
pressed. For the sprint key, it should be 1.0 when released and some
number other than 1.0 when pressed. To keep the acceleration between
them the same, we need to subtract 1.0 from
*m_directions_lerp[dir][value] in the total_ramp_time equation, just for
sprint.