Apply a constant acceleration even when sprinting
The top speed is scaled based on the sprint event value multiplier as
well as the acceleration / deceleration (which is now a constant when
sprinting).

The velocity vs time can be described using a line which has a slope
defined by the acceleration value that is entered in the editor, with
the sprint event value multiplier adjusting that slope (accleration) by
a factor of itself. The sprint velocity vs time line is effectively slid
left/right along the time axis and the intersection of this steeper line
and the "walking" line describes the transition from walking to
sprinting. The same two-segment lines produced when sprinting beyond the
max walk speed is followed in reverse when decelerating.

If sprint was used up to the point of reaching any arbitrary speed, then
the action of releasing the movements keys will cause a
sprint-multiplied deceleration to occur depending on how long the sprint
key was pressed, up to a maximum of
((*m_directions_lerp[dir][value]-1.f)*m_speed)/m_accel seconds.