Do not apply sprint if the target velocity is applying a backwards movement
Previously, this was depending on the applied velocity direction (with
respect to the character's frame of reference). This is not the desired
behavior since doing so would mean that the sprint adjust factor is
still calculated (and can become negative) when still moving forward but
intending to move back. It also previously meant that the counter would
still increment even though the intended movement is backwards.