Use AZ::Vector3's lerp function to linearly interpolate the entire velocity vector rather than its individual components
Adjust the sprint factor based on the character's movement angle with
respect to their own frame of reference.

Scale each axes' positive and negative velocity components independently, such
that it can be used to produce an X/Y graph such as this: https://imgur.com/a/tI5Sdkk
Where the green area is the expected range of values that the velocity
can be within. In this example, the negative Y direction is scaled by
roughly half the rest. Therefore, if an input value of -1.0 in the Y
axis was entered, then it would result in -0.5; if an input value of
-0.5 was entered, then it would result in -0.25, etc.