Add a basic jumping mechanic using CreateCapsuleCastRequest to determine if the player is grounded
Currently, the initial velocity is applied continiously without a -Z
acceleration due to gravity component until either the jump key is
released or the shapecast distance is exceeded (with some logic
preventing you from re-pressing the jump key in mid-air while still in a
"grounded" state).

Next up will be to add a toggle for gravity so that the designer can
choose whether to use this component's gravity or the PhysX Character
Gameplay component's gravity (or no gravity at all).