Use CreateCapsuleOverlapRequest instead of CreateCapsuleCastRequest to check the grounded state
The capsule's pose needed to be rotated by 90 degrees on the Y axis so
that its height is oriented along the Z axis.

The overlap detection capsule is offset in the negative Z direction by
half of the character's collider height, so the offset should be
+(character height)/2 minus how much distance you want to detect beneath
the character. The more you subtract, the higher you can jump (depending
on how long you hold the jump key).