Make it so that the sprint held counter is decremented after a pause time that is determined using the ratio of the time the sprint key was held divided by the max held time and the cooldown time.
For the case where the max held time is longer than the cooldown time, a
constant is used instead. In this case it is hard-coded as an arbitrary
constant of 0.1 times the cooldown time. It is not exposed to the editor
since it is only used in the case where the max held time is longer than
the cooldown time; when the reverse is true an equation is used to
determine how long the pause is (based on how long the key has been
cumulatively pressed). When the max held time is longer than the
cooldown time it is more advantageous to always hold down the sprint key
since that ensures that the average velocity is maximized (if that is
the desire). By adding a pause in this case, it simply deters the desire
to spam the key in an attempt to maintain the sprint from elapsing and
triggering a cooldown (though doing so would mean only achieving an
average velocity around half way between the sprint and walk velocity).

Also make it so that the m_sprint_held_duration delta is multiplied by a
factor of m_sprint_velocity_adjust since left/right strafing has
effectively no sprint applied when just left/right is pressed.

Use HalfPi instead of Pi/2 in the m_sprint_velocity_adjust calculation.