Compute the max jump hold time based on whether or not the apogee of the jump will be inside or outside of the overlap capsule offset that's used in its computation
If the apogee is inside the capsule then the hold time is calculated by
dividing the overlap capsule offset height by the average of the initial
velocity and the velocity at the very edge of the overlap capsule (which
is also computed). In the case where m_jump_held_gravity_factor is 0,
this results in the same calculation as before when the initial jump
held velocity was constant.

Otherwise, the apogee is calculated using one of the simple linear
equations of motion from elementary physics. This way, when the apogee
is reached, the jump key is no longer considered to be held.