Check to see if the angle between the sum of all of the normal vectors which have a steep angle in the grounded check sum to produce a resultant vector whose angle with respect to the +Z unit vector is less than or equal to the specified maximum grounded slope angle
If it is less than the specified maximum slope angle then the character
should be considered to be grounded, and gravity is not accumulated.

If there is an intersection with any individual entity that has an angle
less than the specified maximum grounded slope angle then being grounded
is guaranteed.