Account for the rare instances where the initial jump velocity is not applied by also checking m_z_velocity <= 0.f in the grounded condition
By doing this we're preventing a semi-rare case where m_z_velocity is
mistakenly reset to 0 due to the assumption that just because the
current velocity is <= 0 and that the character is grounded then that
means the previous call to AddVelocityForTick caused the current
velocity to be <= 0. This assumed that AddVelocityForTick did
actually execute as expected and that the current Z velocity is
reflective of what the previous call was, but as it turns out this is
not always the case. By also checking if m_z_velocity is <= 0 we are
verifying whether or not the previous tick had it's Z velocity set as
expected, and if the current velocity on Z is <= 0 but m_z_velocity is
not <= 0 then it would not cause m_z_velocity to be reset to 0 (instead
it would continue to the following conditionals' checks as normal in a
jump).

Also rename m_sprint_time to m_sprint_increment_time so that its name is
more reflective of what it's for.

Add some empty-line print statements at the end of the print debug
blocks so that it's easier to read.