Set m_max_grounded_angle_degrees based on the value returned by GetSlopeLimitDegrees from the PhysX Character Controller component's request bus
Remove the "Max Grounded Slope Angle" entry from the serialize context
and edit context list now that it is automatically obtained from the
PhysX Character Controller component. This makes it so that there is no
longer an entry for this variable in the editor.

Also create getter and setter methods for m_max_grounded_angle_degrees
and expose them to the request bus so the value can still be modified by
scripts and other components.