User a sphere cast rather than a capsule cast for the ground detection, ground close detection, and the head obstruction detection mechanism
Enter a non-zero distance for the shapecast's trajectory. It turns out
that using a capsule cast and giving it a distance of 0 in an attempt to
treat it the same as an overlap does not work as expected, particularly
when intersecting with relatively smaller entities, and more
specifically there were issues found near the perimeter/surface of an
intersecting collider.

By "shooting" out a sphere cast from the top and bottom of the
character, these issues are avoided.

Create a getter and setter for the Z velocity.

Create a getter and setter for the head obstruction sphere cast
distance.