Multiply the left and back values by -1 so that the .inputbindings file does not require that the event value multipliers sign is aligned with the intended directions on the X & Y axis
This will also be useful in the case where the .inputbindings file
contains an entire set of all possible keys that might be used to
reassign the input controls to the First Person Controller component.

Also put in checks for m_sprintDecrementPause == 0.f when reassigning
m_sprintDecrementPause so that when it is being set via a script that it
does not get reassigned.

Also assign m_sprintPrevDecrementPause in the setter method for when
m_sprintDecrementPause gets set.