Make it so that m_sprintHeldDuration will decrement at a rate such that the stamina will get back to its full value when nearly depleted at the same amount of time as if you were to just wait the cooldown time for the case when m_sprintCooldownTime > m_sprintMaxTime
This means that the pause is no longer being used in this condition,
instead it has been documented in the comments as an alternative option
as was done prior to this commit.

Additionally, make it so that when the sprint movement is happening
along the local X axis only that the stamina is still able to regenerate
since sprint is effectively not applied (extremely close to 0).