The newVelocity vector in the LerpVelocity method does not get reassigned in the case of decelerating if it is computed to be greater than the previously applied velocity
This will only really occur for a single tick or so, for example when
lerping from +1 on Y to +1 on X it will decelerate for the first half of
the movement and then switch to acceleration on the second half. At
around the half-way point when it is switch from deceleration to
acceleration there is a momement where the computed velocity for the
deceleration case will actually result in a vector whose length is
greater than the last applied velocity, this is what is being accounted
for. In that case we just use the vector that is computed for the normal
acceleration (not altered by the decel factor).