Improve the sprint accumulated time logic such that the acceleration adjustment factor is only captured once when attempting to stop for the case when instantaneous velocity rotation is disabled
Also improve the accumulation scheme in a way that the number will only
increment when the length of the velocity vector increases, and it
will decrease if the velocity vector decreases, and decreases at an
increased rate if the character is decelerating or experiencing a target
vector that opposes their current velocity by more than 90 degrees.

Unfortunately, there is still some unintended error that can accumulate
with this setup, so m_sprintAccumulateAccelTime is only guaranteed to
properly track the sprint acceleration time in the more simple case. The
error accumulates when the character is rotating.