Add the ability to set whether the character is sprinting via the request bus (script)
Fix a bug where m_sprintValue could still be used during cooldown.

Make it so when sprinting while crouching is enabled that sprinting
actually happens on the way down. Also make it so that if you press
crouch while sprinting and m_sprintWhileCrouched,
m_crouchSprintCausesStanding, and m_crouchPriorityWhenSprintPressed are
true, the character will still move downward until they are fully
crouched and will stand back up if they are sprinting (making the
behavior consistent for when m_sprintWhileCrouched is false but
m_crouchSprintCausesStanding and m_crouchPriorityWhenSprintPressed are
true).