Make it so that when the character's "head" hits a collider the Z velocity is set to 0 and that becomes the apogee of their jump
Allow for non-kinematic (dynamic) rigid bodies to either be ignored or
considered for this check.

Create getters and setters for whether something was hit, the distance
that the sphere cast will check for collisions, the collision group
name, whether non-kinematic rigid bodies are considered, and the
entityIds that are hit (available via the request bus).

Create a notification bus for when something hits the character's head.