Add the ability to change the direction of the jump and gravity "Z" axis via the SetVelocityZPosDirection() method, accessible via the the request bus
Also, separately, add the ability to define a new axis which the sphere
casts move along using SetSphereCastsAxisDirectionPose(), also accessible
via the request bus. This method will also reposition the sphere casts'
pose based on this new orientation (the idea is that they would be
located at the new position of the character's head and feet, once
CharacterControllerComponent::SetUpDirection() is implemented).

Setting both m_velocityZPosDirection and m_sphereCastsAxisDirectionPose
to the same vector causes them to be aligned. One use-case would be to
set them equal to the normal vector of a surface that the character is
standing on, effectively reorienting the character such that their new
"Z" velocity direciton is perpendicular to the surface, along with the
directions that the shapecasts move.

When the ability to also reorient the PhysX Character Controller's
collider is implemented, the entire character can be rotated with what
is provided.