Add the ability to tilt the XY plane with respect to the Z axis using SetVelocityXCrossYDirection()
When m_velocityXCrossYDirection is set equal to the normal vector of a
surface that the ground sphere cast obtained then the character can be
made to have the X&Y components of its velocity coplanar with the
surface that its standing on.

This can be used to remove the stair-step effect when moving down an
incline. Care should be taken to reorient the XY plane after doing so
(e.g. when the GetGroundSceneQueryHits() method returns an empty
vector).

The Z axis is untouched when using SetVelocityXCrossYDirection().