Assign the sphereCastDirection variables directly to m_sphereCastsAxisDirectionPose (or its negative) rather than applying quaternion operations to get it to point in the same direction
Use m_sphereCastsAxisDirectionPose in the steep normals calculation.

Create another vector of SceneQueryHit for the ground close detection
and create a getter method for this new member, called
GetGroundCloseSceneQueryHits().

Account for when m_velocityXCrossYDirection's and
m_velocityZPosDirection's Z component is negative.

Create getter methods for the sum of all of the ground and ground close
detection's normal vectors called GetGroundSumNormalsDirection() and
GetGroundCloseSumNormalsDirection(), accessible via the request bus.

Place the velocity XY plane tilting and new +Z direction code into their
own functions, which has the target velocity passed to it by reference.

Make it so that the gound close detection always happens, regardless of
whether or not m_grounded is true or not (this is useful when
m_groundCloseHits isn't empty but m_groundHits is, e.g. when trying to
get the normal vectors of a near by surface but the character is not
grounded).