Make it so that you can select the sprint velocity scale values on a per-direction basis
Create a new method called CreateScaledVector which takes as its input a
vector and four scale points, forward, back, left, and right and
produces a vector which points in the same direction as the input vector
but is scaled using ellipses in each of the four quadrants (connecting
each of the points together). If the input vector is normalized (has a
magnitude of 1) then each of the direction scale points are used as-is,
if the input vector has a magnitude of anything other than 1 then its
magnitude is multiplied by each of the individual four scale values to
produce a stretched/compressed resultant vector accordingly.