Check whether the entity hit is Dynamic instead of Non-Kinematic for standing and jumping
Since O3DE 2305.0 now delineates between kinematic and simulated bodies,
simply checking if an entity is not kinematic is no longer sufficient.

By checking whether a null pointer is retrieved when attempting to get
the RigidBody of the entity that's hit using the RigidBodyRequestBus we
can determine if it is static or not.

Due to the following change in O3DE:
https://github.com/o3de/o3de/pull/14261