Make it so that the character will crouch back down after performing a jump if "Crouch Jump Causes Standing" is enabled and "Crouch Enable Toggle" is disabled
After performing the jump the character will crouch back down while the
crouch key and jump key continue to be held. The jump key will need to
be released and repressed to reperform another jump.

This behavior is consistent with the case when toggle is enabled. When
toggle is enabled the crouch key needs to be pressed again, but when
toggle is disabled it behaves as though the key is being constantly
polled (pressed over and over until the crouch is actually performed).