Obtain the PhysX Character Controller Component's capsule collider dimensions and maximum slope angle one time (on the first tick)
This is necessary since the PhysX Character Controller Component does
not have a handler when inside the First Person Controller Component's
Activate() method. Even if the First Person Controller Component is set
as a EntityBus handler and overrides OnEntityActivated(), the PhysX
Character Controller Component still does not have an event bus handler
up to that point. Which is why this workaround is needed.