Add an event notification for the apogee of jumps called OnJumpApogeeReached
Make it so that OnStartedFalling only occurs when the previous Z
velocity was 0, indicating that it is a fall rather than the apogee of a
jump.

Make sure m_jumpHeld gets set to false at the appropriate time, when the
jump key is released in the middle of a jump.

The combination of these changes will allow for the creation of a script
that can modify m_jumpFallingGravityFactor if the jump key continues to
be held after the apogee is reached, allowing the rate of decent to be
modified accordingly (e.g. falling slower if the key keeps being held).