Make the crouching motion linear again, instead of using SmoothCriticallyDampedFloat()
Revert this back to being linear because the smoothTime for the
smoothCD function cannot trivially derive its value from m_crouchTime.
It's not as simple as a constant ratio. The smoothTime is the time it
would take for smoothCD to get to its target if the velocity were kept
constant when the function reaches its maximum (see Figure 1.10.4 in
Andrew Kirmse's Game Programming Gems 4).

Keep the SmoothCriticallyDampedFloat() function in case it is desired
for something else.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>