Add constant deceleration to the crouching motion
Make it so that the crouching motion is not just a constant velocity up
and down. Add the ability to enter an initial velocity for both
crouching and standing independently. Make it so that if the user enters
an initial velocity that is too small to get to the m_crouchDistance
within the specified time then a constant velocity is used instead (and
print a warning). Also make it so that if the user enters too large of
an initial velocity then a smaller value is used instead which has the
final velocity go down to zero (and print a warning).

Use the current deltaTime plus the previous deltaTime divide by two in
the ProcessInput() function's various calls.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>