Add the ability to apply impulses to the character
Impulses are in units of N*s or kg*m/s and indicate the initial velocity
that will be applied based on the character's mass. A deceleration is
used to decay the velocity from its initial value down to zero. Expose
all attributes with setters and getters to the EBus.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>