Add an optional capsule shapecast for detecting collisions / hits
The capsule size is set based on a percentage of the PhysX Character
Controller's capsule.

Add setters and getters for the capsule size, collision group, whether
hit by static, dynamic, or both types of rigid bodies, and a vector of
scene query hits. Use a bool to determine if the feature is enabled.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>