Have OnPhysicsTimestepStart() and OnPhysicsTimestepFinish() return the time step
Make the capsule shapecast project in the direction of the velocity (by
the velocity times the timestep).

Track the velocity in a new member variable.

Check for obstructions on the physics timestep.

Add a member variable that sets an additional sweep distance on the
shapecast which is a percentage of the character collider radius.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>