Add the option to slow down the character's movement when going up an incline
Use a new bool called m_movingUpInclineSlowed which sets whether or not
this functionality is enabled. Add a getter and setter method that is
exposed to the behavior context for it to be enabled or disabled via
scripts. Make it so that this bool is selectable from the editor. It is
only able ot be enabled if m_velocityXCrossYTracksNormal is also set to
true.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>