Conserve the lateral (X&Y) velocity when jumping on inclined surfaces, particularly when jumping down them
When m_movingUpInclineSlowed == true and jumping up inclines the lateral velocity is capped based on what it would be limited to when walking on the surface.

Fix an issue where jumping up inclines was properly limiting the component of the lateral (X&Y) velocity in the direction of the incline, but was not preserving the length of X&Y together.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>