Accumulate the velocity from impulses inside ProcessLinearImpulse() rather than in ProcessInput()
Specifically, add the velocity due to impulses to m_applyVelocityZ and
m_applyVelocityXY so that the OnJumpApogeeReached() and
OnStartedFalling() correctly detect the max height when impulses are
applied. Additionally, reset the Z component of the velocity immediately
after adding the initial velocity for the projectile motion equation to
take its course. The X&Y components of the impulse velocity will
continue to sum but decay based on m_impulseConstantDecel and/or
m_impulseLinearDamp.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>