Fix the issue where mulitple jumps can be performed without gravity applied during coyote time
Fix the issue where jumping isn't working during coyote time when
m_applyGravityDuringCoyote == true.

Add a bool which can be used to determine whether the last surface
normal is kept track during the coyote time (when gravity is not
applied). Make it access via the ebus with a getter and setter.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>