Apply the m_maxGroundedAngleDegrees to the PhysX Character Controller using SetSlopeLimitDegrees
m_maxGroundedAngleDegrees is being set to 0.01 greater than the maximum slope
angle that is obtained from the PhysX Character Controller in an attempt
to make it so that the character can walk on all surfaces that are
_exactly_ the maximum slope angle. However, when walking off inclines
that are exactly the maximum slope angle, the xy velocity vector can
flip when tracking surface inclines. This only occurs when the PhysX
Character Controller does not correctly recognize the surface that the
character is standing on to be within the valid angle range, even though
it is at the very limit. This only happens for some inclines at the
limit, depending on how the entity is rotated. So to work around this
bug, the slight increase in angle is applied back to the PhysX Character
Controller.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>