Attempt to fix the issue where clients are persistently getting desynced with the server due to the server running multiple instances for each client thereby broadcasting and listening to multiple notification events (one for each running client)
The solution is to include the EntityId as part of the notification
broadcast and have the listeners check if their EntityId is the same one
that is passed as an argument.

Add network properties for ApplyVelocityXY and ApplyVelocityZ and
synchronize them with the server using the getters and setters. Do the
work inside newly created functions called GetNetworkFPCProperties(...)
and SetNetworkFPCProperties().

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>