Add a new network property that carries the transform that results after doing TryMoveWithVelocity(...)
The non-autonomous version of the entity can report a different
transform/translation than what the autonomous client calculates and
observers, so this may come in handy. However, when retrieving the
network property value, it can be stuck for two consecutive network
ticks (likely due to networking latency), so it may not be as reliable
as just using whatever transform is obtained from the simulation.

Signed-off-by: Eric Kuzmenko <eric.n.kuzmenko@gmail.com>